# Perf: Reduce time taken to parse hive partitions