I am working with a parquet dataset with 53960 parquet files across 29 hive partitions. ie `batch=<0..29>`

Removing all files sharing the same partition reduced the time taken to parse the dataset by 88%.

Before: getting_metadata/Load partition data for large dataset
                        time:   [647.28 ms 656.55 ms 669.53 ms]
                        change: [-1.1985% +0.5724% +2.8272%] (p = 0.63 > 0.10)
                        No change in performance detected.

After: getting_metadata/Load partition data for large dataset
                        time:   [54.929 ms 55.434 ms 56.046 ms]
                        change: [-89.091% -88.982% -88.851%] (p = 0.00 < 0.10)
                        Performance has improved.